action widgetをpull indicatorに変更
