--- test/multi_pull_test.dart
@@ -1,6 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-import 'package:multi_pull/multi_pull.dart';
-
-void main() {
-
-}
